requirements.txt
PyInquirer~=1.0.3
asyncpg~=0.27.0
fastapi~=0.96.0
grpcio-tools~=1.54.0
psycopg2-binary~=2.9.6
pydantic~=1.10.7
rich~=13.3.5
uvicorn~=0.22.0
wrapt~=1.15.0
